Update devcontainer.json
Use postStartCommand instead of postCreate.

For some reason, postCreate does not seem to run after creation when Full Rebuild is run within a codespace. (Although it does seem to work as intended when created directly from within the github app with 'create a codespace from this template')   Presumably this is a bug with codespaces or devcontainers. postStartCommand avoids the problem (though is slighly hacky)